Upcoming Release Changes (#56)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @theguild/federation-composition@0.10.1

### Patch Changes

-   [#55](https://github.com/the-guild-org/federation/pull/55)

[`5c4431d`](https://github.com/the-guild-org/federation/commit/5c4431da9ce343d3bb7bf9db30b6e2ceb026e2e3)
    Thanks [@n1ru4l](https://github.com/n1ru4l)! - fix esm support

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>