Upcoming Release Changes (#61)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @theguild/federation-composition@0.11.2

### Patch Changes

- [#60](https://github.com/the-guild-org/federation/pull/60)
[`2f7fef1`](https://github.com/the-guild-org/federation/commit/2f7fef10409a25f8366182448a48e72d5451abf9)
Thanks [@kamilkisiela](https://github.com/kamilkisiela)! - Normalize
enum values to be printed as enum values in Supergraph SDL, even if the
user's subgraph schema has them as strings

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>