Upcoming Release Changes (#77)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @theguild/federation-composition@0.14.2

### Patch Changes

- [#76](https://github.com/the-guild-org/federation/pull/76)
[`a3cb724`](https://github.com/the-guild-org/federation/commit/a3cb7241dc1e07e8f8c0f2e0c2908a68c55f66d9)
Thanks [@kamilkisiela](https://github.com/kamilkisiela)! - Fix a missing
`@join__field` on a query field where `@override` is used, but not in
all subgraphs.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>